requirements.txt
requests==2.18.4
bs4==0.0.1
numpy==1.14.3
tensorflow==1.4.0
keras==2.1.6
h5py==2.7.1
elasticsearch==5.5.2
python-dotenv==0.8.2
tornado==5.0.2
Tornado-JSON==1.3.2
